# apps/web/openapi/axon.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Axon REST API",
    "description": "Dedicated REST routes for Axon discovery, RAG, crawl, ingest, and watch workflows.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "6.2.2"
  },
  "paths": {
...
    "/v1/ask": {
      "post": {
        "tags": [
          "rag"
        ],
        "operationId": "v1_ask",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestAskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "RAG answer",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid ask request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated token lacks Axon access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Ask request exceeds limits",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "Upstream vector or LLM service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "504": {
            "description": "Upstream request timed out",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "oauth2": [
              "axon:read"
            ]
          },
          {
            "oauth2": [
              "axon:write"
            ]
          }
        ]
      }
    },
    "/v1/ask/stream": {
      "post": {
        "tags": [
          "rag"
        ],
        "operationId": "v1_ask_stream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestAskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "RAG answer streamed as server-sent events",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ask request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated token lacks Axon access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Ask request exceeds limits",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "oauth2": [
              "axon:read"
            ]
          },
          {
            "oauth2": [
              "axon:write"
            ]
          }
        ]
      }
    },
...
    "/v1/extract": {
      "get": {
        "tags": [
          "jobs"
        ],
        "operationId": "list_extract_jobs",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Extract jobs",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated token lacks Axon access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "oauth2": [
              "axon:read"
            ]
          },
          {
            "oauth2": [
              "axon:write"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "jobs"
        ],
        "operationId": "start_extract",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestExtractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Extract job accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedJob"
                }
              }
            }
          },
          "400": {
            "description": "Invalid extract request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated token lacks Axon access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "502": {
            "description": "Upstream extract service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "oauth2": [
              "axon:read"
            ]
          },
          {
            "oauth2": [
              "axon:write"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "operationId": "clear_extract_jobs",
        "responses": {
          "200": {
            "description": "Extract clear result",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated token lacks Axon access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "oauth2": [
              "axon:read"
            ]
          },
          {
            "oauth2": [
              "axon:write"
            ]
          }
        ]
      }
    },
    "/v1/extract/cleanup": {
      "post": {
        "tags": [
          "jobs"
        ],
        "operationId": "cleanup_extract_jobs",
        "responses": {
          "200": {
            "description": "Extract cleanup result",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated token lacks Axon access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "oauth2": [
              "axon:read"
            ]
          },
          {
            "oauth2": [
              "axon:write"
            ]
          }
        ]
      }
    },
    "/v1/extract/recover": {
      "post": {
        "tags": [
          "jobs"
        ],
        "operationId": "recover_extract_jobs",
        "responses": {
          "200": {
            "description": "Extract recovery result",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated token lacks Axon access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "oauth2": [
              "axon:read"
            ]
          },
          {
            "oauth2": [
              "axon:write"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptedJob": {
        "type": "object",
        "required": [
          "job_id",
          "status",
          "status_url"
        ],
        "properties": {
          "job_id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "status_url": {
            "type": "string"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "kind",
          "message"
        ],
        "properties": {
          "diagnostics": {
            "type": "object"
          },
          "kind": {
            "$ref": "#/components/schemas/ErrorKind"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "ok",
          "contract_version",
          "error",
          "request_id",
          "trace"
        ],
        "properties": {
          "contract_version": {
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "ok": {
            "type": "boolean"
          },
          "request_id": {
            "type": "string"
          },
          "trace": {
            "$ref": "#/components/schemas/TraceContext"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or static token"
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/authorize",
            "tokenUrl": "/token",
            "scopes": {
              "axon:read": "Authenticated Axon REST access",
              "axon:write": "Authenticated Axon REST access"
            }
          }
        }
      }
    }
  },
...
}